+++ pkg/casheerapi/total.go
@@ -4,12 +4,12 @@ const TotalType = "total"
 
 type TotalData struct {
 	ResourceID
-	Month          int     `json:"month"`
-	Year           int     `json:"year"`
-	ExpectedIncome float32 `json:"expected_income"`
-	RunningIncome  float32 `json:"running_income"`
-	ExpectedTotal  float32 `json:"expected_total"`
-	RunningTotal   float32 `json:"running_total"`
+	Month          int   `json:"month"`
+	Year           int   `json:"year"`
+	ExpectedIncome int64 `json:"expected_income"`
+	RunningIncome  int64 `json:"running_income"`
+	ExpectedTotal  int64 `json:"expected_total"`
+	RunningTotal   int64 `json:"running_total"`
 }
 
 type GetTotalParams struct {
